requirements.txt
pip~=25.0
pillow~=11.1.0